Restructuring things a bit.
